COR-288: Fixed also the OrganizationAuthenticatorImpl
