+++ exo.core.component.organization.api/src/main/java/org/exoplatform/services/organization/impl/mock/DummyOrganizationService.java
@@ -123,6 +123,11 @@ public class DummyOrganizationService extends BaseOrganizationService
          Collection memberships = new ArrayList();
          if ("root".equals(userName) || "john".equals(userName) || "admin".equals(userName))
          {
+            MembershipImpl member = new MembershipImpl();
+            member.setMembershipType("member");
+            member.setUserName(userName);
+            member.setGroupId(GROUPID_ADMINISTRATORS);
+            memberships.add(member);
             MembershipImpl admin = new MembershipImpl();
             admin.setMembershipType("*");
             admin.setUserName(userName);
