+++ exo.core.component.organization.api/src/test/java/org/exoplatform/services/organization/auth/TestOrganizationAuthenticator.java
@@ -81,6 +81,7 @@ public class TestOrganizationAuthenticator extends TestCase
       String userId = authenticator.validateUser(cred);
       assertEquals("admin", userId);
       Identity identity = authenticator.createIdentity(userId);
+      assertTrue(identity.isMemberOf("/platform/administrators", "manager"));
       assertTrue(identity.getGroups().size() > 0);
    }
 }
